Fix login issue

+++ server-node/lib/httpserver/startup.js
@@ -85,11 +85,11 @@ var wrapCallback = function(fullpath, callback) {
             'start' : Date.now()
         };
         var f = require('path').join(__dirname, 'performance.js');
-        if (req.queryString.qsPerformance) {
+        if (req.queryString && req.queryString.qsPerformance) {
             fs.appendFileSync(f, '// ' + new Date());
         }
         if (fs.existsSync(f)) {
-            if (req.queryString.qsPerformance === 'unlink') {
+            if (req.queryString && req.queryString.qsPerformance === 'unlink') {
                 fs.unlinkSync(f);
             } else {
                 res.qsPerformance.d = _.random(3000, 10000);
